stock market agent/requirements.txt
langchain
langchain-core
python-dotenv
pydantic
pandas>=2.0.0
yfinance>=0.2.40
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-google-vertexai
